Added all my code
This is the current state of robotop but its a couple features behind what
I had. I think all I'm missing is the weapon class which wasn't actually being
used at the time anyways, but it worked. I had a Lua.cpp file or something
of that nature. And I'm missing some new tiles I had. I'm just glad I found
this build and that it's almost completely up to date.